Refactor: FSD 원칙에 따라 파일 구조 및 네이밍 리팩터링 (#15)
