+++ src/entities/group/api/get-groups.ts
