+++ src/entities/group/api/group.query.ts
