+++ src/entities/group/index.ts
@@ -1,1 +1,3 @@
-export { GroupCard } from './ui/group-card'
+export * from './ui/group-card'
+export * from './ui/group-plan'
+export * from './ui/group-member-list'
