+++ src/features/group/index.ts
@@ -1,0 +1,2 @@
+export * from './switch-group-info-tabs/ui/switch-group-info-tabs'
+export * from './switch-group-info-tabs/model/types'
