+++ src/features/group/switch-group-info-tabs/model/types.ts
@@ -1,0 +1,4 @@
+export type SwitchGroupInfoTabsProps = {
+    activeTab: "info" | "schedule";
+    onTabChange: (tab: "info" | "schedule") => void;
+};
