+++ src/features/search/search-groups/ui/search-bar.tsx
