--- src/widgets/group-detail/index.ts
@@ -1,5 +1,0 @@
-export { GroupDetailHeader } from './ui/group-detail-header';
-export { GroupInfoTabs } from './ui/group-info-tabs';
-export { MembersList } from './ui/members-list';
-export { GroupPlan } from './ui/group-plan';
-export { useGroup } from './lib/use-group';
