+++ src/widgets/group-details/index.ts
@@ -1,0 +1,3 @@
+export { GroupInfo } from './ui/group-info';
+export { useGroup } from './lib/use-group';
+export * from './lib/use-group';
