+++ src/widgets/group-details/lib/use-group.ts
