+++ src/widgets/group-details/lib/use-schedules.ts
